requirements.txt
streamlit==1.38.0
youtube_transcript_api==0.6.2
langchain==0.3.0
langchain-huggingface==0.1.0
langchain-google-genai==2.0.0
langchain-community==0.3.0
langchain-core==0.3.56
faiss-cpu==1.8.0
python-dotenv==1.0.1
markdown==3.7
torch==2.4.1
transformers==4.44.2
sentence-transformers==3.1.1




